fixed loggings, fixed int( ) conversions

+++ src/calc/calculate_phase1.py
@@ -66,11 +66,10 @@ class CalculatePhase1(CalculatePhaseBase):
             rl1.ratio1 = rl1.ratio
 
         # total reward amount needs to be diminished at the same rate total balance diminishes
+        # TODO: Why is the Decimal class created?
         new_total_amnt_multiplier = new_total_balance / total_balance
-        new_total_amount = int(
-            Decimal(total_amount * new_total_amnt_multiplier).to_integral_value(
-                rounding=ROUND_HALF_DOWN
-            )
-        )
+        new_total_amount = Decimal(
+            total_amount * new_total_amnt_multiplier
+        ).to_integral_value(rounding=ROUND_HALF_DOWN)
 
-        return rewards, new_total_amount
+        return rewards, int(new_total_amount)
